fix: keep SafeChat backend internal
