+++ .github/workflows/ci.yml
@@ -110,7 +110,6 @@ jobs:
                 filename.startsWith('Formula/') ||
                 filename.startsWith('build/') ||
                 filename.startsWith('internal/msgcrypto/') ||
-                filename.startsWith('internal/app/safechat_command') ||
                 filename.startsWith('third_party/safechat-go-sdk/') ||
                 filename.startsWith('third_party/runtimepayload/') ||
                 filename === '.goreleaser.yaml' ||
@@ -1756,8 +1755,8 @@ jobs:
       - name: Enable MinGW for SafeChat CGO
         shell: pwsh
         run: |
-          "C:\msys64\mingw64\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
-          & "C:\msys64\mingw64\bin\gcc.exe" --version
+          "C:\mingw64\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
+          & "C:\mingw64\bin\gcc.exe" --version
 
       - name: Build Windows CLI
         run: go build -o dws.exe ./cmd
@@ -1835,8 +1834,8 @@ jobs:
       - name: Enable MinGW for SafeChat CGO
         shell: pwsh
         run: |
-          "C:\msys64\mingw64\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
-          & "C:\msys64\mingw64\bin\gcc.exe" --version
+          "C:\mingw64\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
+          & "C:\mingw64\bin\gcc.exe" --version
 
       - name: Resolve authoritative coverage base
         shell: bash
